Create first VMs for DNS
Finally! Going with the naming scheme of dog breeds for the machines.

Alongside this recursive dns resolver I have one running on a Pi (which
also used to be authoritative). I've replaced bind on it with unbound
manually, as I'll bring it into ansible later, and am going to have just
a single authoritative which both recursors point at - I only care about
redundancy for the recursion for home use.