Add MySQL server and provider
VM provisioned by third party ansible role and has a terraform user and
provider to manage the databases/users going forward. The pre_tasks is
because of a bug with Debian 10 where ansible tries to use the wrong
pymysql module.

Plan to set up replication at some point in the future which the role
can do.