Rebuild k8s cluster with calico as cni
Does some nicer routing using BGP (which my router supports) and no
overlay networking, as well as policies. Easier to just rebuild the
whole cluster so also replaced IP addresses.

The BGP routing negates the need for a load balancer so remove metallb
and get ingress to use IP which will be advertised.